Move the HeroBanner component out of __common__
