+++ components/HeroBanner/HeroBanner.css
@@ -1,6 +1,35 @@
 .HeroBanner {
   align-items: center;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  color: var(--white);
   display: flex;
-  margin: 1rem auto;
+  flex-direction: column;
+  min-height: 65vh;
+  justify-content: center;
+  text-shadow: #111111 0 0 15px;
+  width: 100%;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  height: auto;
   width: 90%;
+  align-self: center;
+  margin-top: calc(3rem + var(--navBarHeight) + var(--navBarTopGutterHeight));
+}
+
+.underline {
+  border-bottom: 4px solid var(--primary);
+  margin-bottom: 1rem;
+  text-align: center;
+}
+
+.fullViewHeight {
+  min-height: 100vh;
+  width: auto;
 }
